refactor: move autcomplete and archivedautocomplete component to config goal's component folder
