+++ src/components/GoalsComponents/GoalConfigModal/components/AutoComplete.scss
